refactor: reorganize styles by creating separate files for FavoritesRow, ProductsCard, and SearchResults; remove deprecated home-components styles
